+++ library-navigational-compat/src/main/java/universum/studios/android/transition/NavigationalTransitionCompat.java
@@ -74,7 +74,7 @@ public class NavigationalTransitionCompat extends BaseNavigationalTransition<Nav
 	 *                                  whenever starting this transition.
 	 * @see #createIntent(Activity)
 	 */
-	public NavigationalTransitionCompat(@NonNull Class<? extends Activity> classOfTransitionActivity) {
+	public NavigationalTransitionCompat(@NonNull final Class<? extends Activity> classOfTransitionActivity) {
 		super(classOfTransitionActivity);
 	}
 
@@ -83,8 +83,9 @@ public class NavigationalTransitionCompat extends BaseNavigationalTransition<Nav
 	 */
 
 	/**
-	 * Starts this navigational transition using the given <var>caller</var> fragment with all
-	 * transitions specified for this navigational transition.
+	 * Starts this navigational transition using the given <var>caller</var> fragment with transitions
+	 * and shared elements that are configured for the caller's parent activity via
+	 * {@link #configureOutgoingTransitions(Activity)}.
 	 * <p>
 	 * <b>Note</b>, that unlike {@link ##start(Activity)} this cannot be used to start new activity
 	 * with shared elements presented. For that purpose use {@link #start(Activity)} instead.
@@ -94,10 +95,9 @@ public class NavigationalTransitionCompat extends BaseNavigationalTransition<Nav
 	 * @see #configureTransitionsOverlapping(Activity)
 	 * @see #configureTransitions(Activity)
 	 */
-	public void start(@NonNull Fragment caller) {
+	public void start(@NonNull final Fragment caller) {
 		final Activity activity = caller.getActivity();
-		configureTransitionsOverlapping(activity);
-		configureTransitions(activity);
+		configureOutgoingTransitions(activity);
 		onStart(caller);
 	}
 
@@ -114,7 +114,7 @@ public class NavigationalTransitionCompat extends BaseNavigationalTransition<Nav
 	 */
 	@SuppressLint("NewApi")
 	@SuppressWarnings("ConstantConditions")
-	protected void onStart(@NonNull Fragment caller) {
+	protected void onStart(@NonNull final Fragment caller) {
 		final Activity activity = caller.getActivity();
 		final Intent intent = createIntent(activity);
 		if (MATERIAL_SUPPORT) {
@@ -134,7 +134,7 @@ public class NavigationalTransitionCompat extends BaseNavigationalTransition<Nav
 	 *               should be started.
 	 * @see #start(Fragment)
 	 */
-	public void finish(@NonNull Fragment caller) {
+	public void finish(@NonNull final Fragment caller) {
 		onFinish(caller);
 	}
 
@@ -147,7 +147,7 @@ public class NavigationalTransitionCompat extends BaseNavigationalTransition<Nav
 	 * @param caller The fragment that requested finish of its host activity via this navigational transition.
 	 * @see #onStart(Fragment)
 	 */
-	protected void onFinish(@NonNull Fragment caller) {
+	protected void onFinish(@NonNull final Fragment caller) {
 		onFinish(caller.getActivity());
 	}